Bring the things together in device.kt
